requirements.txt
Flask
Pillow
PyPDF2
werkzeug
gunicorn
